32-bit compatibility changes - Add long long versions for database_result and config - Use long long for timestamps and byte counts to ensure we work   on 32-bit systems (mainly targetting raspbian 32-bit)

+++ export.hpp
@@ -46,8 +46,8 @@ private:
 
     //details of current job:
     std::atomic<long> id;
-    long starttime;
-    long endtime;
+    long long starttime;
+    long long endtime;
     long camera;
     std::string path;
     long progress;
